requirements.txt
amply
datapackage
flatten_dict
importlib_resources; python_version<'3.7'
networkx
pandas
pandas_datapackage_reader
pydot
pyyaml
xlrd
